+++ models/safetensors/safetensors_test.go
@@ -226,7 +226,7 @@ func TestIterTensors(t *testing.T) {
 	m, err := New(repo)
 	require.NoError(t, err)
 	count := 0
-	for tensorAndName, err := range m.IterTensors() {
+	for tensorAndName, err := range m.IterTensors(nil) {
 		require.NoError(t, err)
 		wantShapeStr, found := allMiniVariablesToShape[tensorAndName.Name]
 		require.True(t, found, "tensor %q not expected", tensorAndName.Name)
@@ -239,7 +239,7 @@ func TestIterTensors(t *testing.T) {
 func TestIterTensorsFromRepo(t *testing.T) {
 	repo := hub.New("sentence-transformers/all-MiniLM-L6-v2")
 	count := 0
-	for tensorAndName, err := range IterTensorsFromRepo(repo) {
+	for tensorAndName, err := range IterTensorsFromRepo(nil, repo) {
 		require.NoError(t, err)
 		wantShapeStr, found := allMiniVariablesToShape[tensorAndName.Name]
 		require.True(t, found, "tensor %q not expected", tensorAndName.Name)